[REFACTOR] 파일 위치 변경, 파일명, 폴더명 camel case에 맞게 리팩토링 (#246)
* [FIX] 로그인이 유지 되지 않는 오류 해결 (#243)

* [ADD] 불필요한 Public Url 제거

* #239 [FEAT] OptionalLogin 로직 구현 완료

* [FIX] 로그인 페이지로 이동 button에 Link href 추가 (#244)

* [FIX]모달 path params query params로 변경 (#242)

* [DELETE] 기존 path params 모달 페이지 제거

* [TYPES] PAGE_LIST 상수 readonly

* [FIX] useOpenModal hooks 수정

query params로 변경

* [FEAT] 모달 사용 페이지에서 useRouteModal로 page 정보 전달

* [REFACTOR] utils 내부에 있는 css 파일 styles 내부로 이동

* [REFACTOR] AtomModal.ts lowerCamelCase로 변경

* [REFACTOR] lower camel case 폴더명 camel case로 변경

---------
