+++ pages/setting/profile.tsx
@@ -1,5 +1,5 @@
 import Layout from 'components/common/Layout';
-import { UserProfileCard } from 'components/setting/profile/@index';
+import { UserProfileCard } from 'components/setting/Profile/@index';
 
 const UserProfilePage = () => {
   return (
