+++ pages/social/scrap.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
 import SocialLayout from 'components/social/common/SocialLayout';
-import ScrapList from 'components/social/scrap/ScrapList';
+import ScrapList from 'components/social/Scrap/ScrapList';
 import SideBar from 'components/social/feed/SideBar';
 
 const ScrapPage = () => {
